pyproject.toml
[project]
name = "fastapi-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "fastapi[standard]>=0.116.1",
    "python-dotenv>=1.1.1",
    "pydantic-settings>=2.0.0",
    "ruff>=0.12.7",
    "scalar-fastapi>=1.2.3",
    "sqlmodel>=0.0.24",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I"   # Enable isort rules
]
